# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Authn Platform API
  description: Enterprise Identity Engine REST API specification for authentication, session management, organization multi-tenancy,
    RBAC, SAML, OAuth2/OIDC, Webhooks, API key administration, and account recovery.
  version: 2.0.0
  contact:
    name: Authn Platform Engineering
    url: https://github.com/hanan-bhatti/authn-2.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
- url: http://localhost:8080
  description: Local Development Engine
- url: https://api.staging.authn.dev
  description: Staging Environment
- url: https://api.authn.dev
  description: Production Environment
paths:
...
  /v1/client/auth/2fa/totp/enroll:
    post:
      summary: Enroll TOTP 2FA
      description: Generates a new TOTP secret seed and QR provisioning URI for the authenticated user and stores it in pending
        state.
      operationId: enrollTOTP
      tags:
      - Multi-Factor Authentication
      security:
      - PublishableKey: []
        BearerAuth: []
      responses:
        '200':
          description: TOTP secret and provisioning URI successfully generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TOTPEnrollResponse'
        '401':
          description: Unauthorized session or missing publishable key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/client/auth/2fa/totp/confirm:
    post:
      summary: Confirm TOTP 2FA Enrollment
      description: Validates a 6-digit TOTP code against the pending secret, enables TOTP 2FA, and optionally returns generated
        backup recovery codes.
      operationId: confirmTOTP
      tags:
      - Multi-Factor Authentication
      security:
      - PublishableKey: []
        BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TOTPConfirmRequest'
      responses:
        '200':
          description: TOTP enrollment confirmed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmTOTPResponse'
        '400':
          description: Missing code or invalid TOTP verification code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/client/auth/2fa/totp/disable:
    post:
      summary: Disable TOTP 2FA
      description: Re-verifies user password, removes TOTP 2FA, and revokes all active sessions for security.
      operationId: disableTOTP
      tags:
      - Multi-Factor Authentication
      security:
      - PublishableKey: []
        BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TOTPDisableRequest'
      responses:
        '200':
          description: TOTP 2FA disabled and sessions revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisableTOTPResponse'
        '400':
          description: Password missing or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid password confirmation or unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (e.g. 2FA is mandatory for admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/client/auth/2fa/totp/verify:
    post:
      summary: Verify TOTP Code
      description: Validates a 6-digit TOTP code during a login 2FA challenge (authorized by mfa_token) or within an active
        authenticated session.
      operationId: verifyTOTP
      tags:
      - Multi-Factor Authentication
      security:
      - PublishableKey: []
      parameters:
      - in: header
        name: X-Authn-Client-Type
        required: false
        schema:
          type: string
          enum:
          - web
          - native
          - mobile
          default: web
        description: Client platform type. Controls whether refresh token is returned in response body or HttpOnly cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TOTPVerifyRequest'
      responses:
        '200':
          description: Verification successful. Returns AuthResponse for login challenge or validity status for session step-up.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyTOTPResponse'
        '400':
          description: Missing code or invalid MFA code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Session expired, invalid token, or unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/client/auth/2fa/sms/enroll:
    post:
      summary: Enroll SMS 2FA
      description: Sends a 6-digit OTP verification code via SMS to the provided phone number to begin SMS 2FA enrollment.
      operationId: enrollSMS
      tags:
      - Multi-Factor Authentication
      security:
      - PublishableKey: []
        BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SMSEnrollRequest'
      responses:
        '200':
          description: Verification code sent via SMS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SMSEnrollResponse'
        '400':
          description: Missing phone number or validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded for SMS requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/client/auth/2fa/sms/confirm:
    post:
      summary: Confirm SMS 2FA Enrollment
      description: Validates the 6-digit SMS OTP code and completes SMS 2FA enrollment.
      operationId: confirmSMS
      tags:
      - Multi-Factor Authentication
      security:
      - PublishableKey: []
        BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SMSConfirmRequest'
      responses:
        '200':
          description: SMS enrollment confirmed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmTOTPResponse'
        '400':
          description: Missing code or invalid/expired SMS OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/client/auth/2fa/sms/disable:
    delete:
      summary: Disable SMS 2FA
      description: Re-verifies user password, disables SMS 2FA, and clears registered phone number.
      operationId: disableSMS
      tags:
      - Multi-Factor Authentication
      security:
      - PublishableKey: []
        BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SMSDisableRequest'
      responses:
        '200':
          description: SMS 2FA disabled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SMSDisableResponse'
        '400':
          description: Password missing or disable failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized or invalid password confirmation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/client/auth/2fa/recovery-codes/regenerate:
    post:
      summary: Regenerate Backup Recovery Codes
      description: Requires password step-up verification, invalidates old recovery codes, and issues 16 new codes.
      operationId: regenerateRecoveryCodes
      tags:
      - Multi-Factor Authentication
      security:
      - PublishableKey: []
        BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoveryCodesRegenerateRequest'
      responses:
        '200':
          description: Recovery codes successfully regenerated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryCodesRegenerateResponse'
        '400':
          description: Password missing or regeneration failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid password confirmation or unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/client/auth/2fa/recovery-codes/status:
    get:
      summary: Get Recovery Codes Status
      description: Returns remaining count of unused backup recovery codes for the authenticated user without exposing code
        values.
      operationId: getRecoveryCodesStatus
      tags:
      - Multi-Factor Authentication
      security:
      - PublishableKey: []
        BearerAuth: []
      responses:
        '200':
          description: Recovery codes status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryCodesStatusResponse'
        '401':
          description: Unauthorized session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    PublishableKey:
      type: apiKey
      name: X-Authn-Publishable-Key
      in: header
      description: Client publishable key identifying application context
    SecretKey:
      type: apiKey
      name: X-Authn-Secret-Key
      in: header
      description: Backend secret key for tenant admin API calls
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Client user or console admin session JWT
  schemas:
...
    UserDTO:
      type: object
      description: Public user profile payload
      required:
      - id
      - email
      - email_verified
      - status
      - created_at
      properties:
        id:
          type: string
          example: usr_1a2b3c
        email:
          type: string
          format: email
          example: user@example.com
        email_verified:
          type: boolean
          example: true
        name:
          type: string
          example: Alex Smith
        status:
          type: string
          example: active
        created_at:
          type: string
          format: date-time
          example: 2026-08-01 12:00:00+00:00
    PolicyWarningDTO:
      type: object
      properties:
        requires_password_upgrade:
          type: boolean
        requires_email_verification:
          type: boolean
        missing_criteria:
          type: array
          items:
            type: string
    AuthResponse:
      type: object
      required:
      - user
      properties:
        user:
          $ref: '#/components/schemas/UserDTO'
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          example: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
        policy_warning:
          $ref: '#/components/schemas/PolicyWarningDTO'
        mfa_required:
          type: boolean
        mfa_token:
          type: string
        methods:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required:
      - error
      - code
      properties:
        error:
          type: string
          example: invalid two-factor authentication code
        code:
          type: string
          example: invalid_mfa_code
    TOTPEnrollResponse:
      type: object
      required:
      - secret
      - uri
      properties:
        secret:
          type: string
          example: JBSWY3DPEHPK3PXP
          description: Base32 TOTP secret seed
        uri:
          type: string
          example: otpauth://totp/Authn:user@example.com?secret=JBSWY3DPEHPK3PXP&issuer=Authn
          description: Provisioning URI for QR code generation
    TOTPConfirmRequest:
      type: object
      required:
      - code
      properties:
        code:
          type: string
          example: '123456'
    ConfirmTOTPResponse:
      type: object
      required:
      - message
      - recovery_codes_created
      properties:
        message:
          type: string
          example: TOTP 2FA enabled successfully
        recovery_codes:
          type: array
          items:
            type: string
          description: Freshly generated recovery codes (returned only upon initial setup/regeneration)
        recovery_codes_created:
          type: boolean
    TOTPDisableRequest:
      type: object
      required:
      - password
      properties:
        password:
          type: string
          example: SuperSecret123!
    DisableTOTPResponse:
      type: object
      required:
      - message
      properties:
        message:
          type: string
          example: 2FA TOTP disabled. All active sessions have been revoked for security.
    TOTPVerifyRequest:
      type: object
      required:
      - code
      properties:
        code:
          type: string
          example: '123456'
        mfa_token:
          type: string
          description: MFA challenge token issued during password login
        method:
          type: string
          description: MFA verification method (totp, sms, recovery_code)
    VerifyTOTPResponse:
      oneOf:
      - $ref: '#/components/schemas/AuthResponse'
      - type: object
        required:
        - valid
        properties:
          valid:
            type: boolean
            example: true
    SMSEnrollRequest:
      type: object
      required:
      - phone_number
      properties:
        phone_number:
          type: string
          example: '+12025550199'
    SMSEnrollResponse:
      type: object
      required:
      - message
      - expires_in_seconds
      properties:
        message:
          type: string
          example: OTP verification code sent via SMS
        expires_in_seconds:
          type: integer
          example: 300
    SMSConfirmRequest:
      type: object
      required:
      - code
      properties:
        code:
          type: string
          example: '482910'
    SMSDisableRequest:
      type: object
      required:
      - password
      properties:
        password:
          type: string
          example: SuperSecret123!
    SMSDisableResponse:
      type: object
      required:
      - message
      properties:
        message:
          type: string
          example: SMS 2FA disabled successfully
    RecoveryCodesRegenerateRequest:
      type: object
      required:
      - password
      properties:
        password:
          type: string
          example: SuperSecret123!
    RecoveryCodesRegenerateResponse:
      type: object
      required:
      - message
      - recovery_codes
      properties:
        message:
          type: string
          example: recovery codes successfully regenerated
        recovery_codes:
          type: array
          items:
            type: string
    RecoveryCodesStatusResponse:
      type: object
      required:
      - remaining_count
      - total_count
      - has_recovery_codes
      properties:
        remaining_count:
          type: integer
          example: 16
        total_count:
          type: integer
          example: 16
        has_recovery_codes:
          type: boolean
          example: true
...